Events updated and example level layout
Events now work but don't passthrough any data, if really needed we can make a new delegate only event which can store data to be read by any event listeners I think. I've given an example guide on how to layout events in Events.cs.
Also the level in this build is a good example of a boss room size with a very short start of level area before which should be expanded a bit based on your level.